refactor(product) : 상품 패키지 분리
